requirements.txt
loguru
flask
psutil
PrettyTable
python-socketio==5.11.2
websockets==12.0
watchdog
click
